evaluation/2D_Information/VisualPerceptionToken/requirements.txt
accelerate==0.34.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==2.4.1
attrs==25.3.0
av==15.1.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
datasets==2.21.0
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.17.5
dill==0.3.8
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
fire==0.7.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.6.1
gmpy2==2.1.2
gradio==5.45.0
gradio_client==1.13.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.5
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
llamafactory==0.9.1.dev0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==1.26.4
nvidia-ml-py==13.580.82
orjson==3.11.3
packaging==24.1
pandas==2.3.2
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
ruff==0.13.0
safehttpx==0.1.6
safetensors==0.6.2
scipy==1.16.2
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==75.1.0
shellingham==1.5.4
shtab==1.7.2
six==1.16.0
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.2
starlette==0.47.3
sympy==1.14.0
termcolor==3.1.0
tiktoken==0.11.0
tokenizers==0.20.3
tomlkit==0.13.3
torch==2.4.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.2
triton==3.0.0
trl==0.9.6
typeguard==4.4.4
typer==0.17.4
typing_extensions==4.15.0
typing-inspection==0.4.1
tyro==0.9.31
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.44.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.20.2
